+++ argilla-server/tests/unit/api/v1/test_datasets.py
@@ -326,6 +326,7 @@ class TestSuiteDatasets:
                     "settings": {
                         "type": "rating",
                         "options": [
+                            {"value": 0},
                             {"value": 1},
                             {"value": 2},
                             {"value": 3},