+++ argilla/CHANGELOG.md
@@ -16,8 +16,13 @@ These are the section headers that we use:
 
 ## [Unreleased]()
 
+### Added
+
+- Added support for rating questions to include `0` as a valid value. ([#4860](https://github.com/argilla-io/argilla/pull/4860))
+
 ### Fixed
 
+- Fix issue when record does not have vectors related ([#4856](https://github.com/argilla-io/argilla/pull/4856))
 - Fix issue on character level ([#4836](https://github.com/argilla-io/argilla/pull/4836))
 
 ## [1.28.0](https://github.com/argilla-io/argilla/compare/v1.27.0...v1.28.0)